Rename parser package to parsers
This is a workaround that fixes: "ModuleNotFoundError: No module named 'parser.qualcommparser'; 'parser' is not a package".
parser in my installation is a built-in module and since I used Python for Windows without any further modifications other people might have this problem.
