+++ parsers/qualcomm_diagcmd.py
